Recognize posctr and ntcctr in mapping file as controls (v1.0.4)
When a line of the mapping file contains "posctr", count the
corresponding sample as a positive control. When a line of the
mapping file contains "ntcctr", count the corresponding sample as a
negative control.

Also, the directories devel_versions and old_versions were dropped
because we now use git exclusively for version control.


Former-commit-id: 889b4f5f1147b57c90d2742e830ed53a626e7da1