Rename ASVs in fasta, abundance tables, and classification keys (v2.1.1) (v2.1.1)
Patch implemented by executing a python script after classifying
ASVs. The script will crash if the sequence of ASVs differs in any
of the files that need to be processed. (In FASTA and
classification keys, ASVs are read top-to-bottom; in abundance
tables, ASVs are read left-to-right.)

Squash-merged from patch/rename_asvs