In part1 check file lengths for sanity (v2.1.2)
Check file lenghts after barcode extraction, splitting reads
by samples (already implemented), and tagcleaning. After each
of the above steps, the pipeline waits for the expected
files to appear with the expected number of lines. Meanwhile,
the qsub error logs are checked for indications of error. This
is a quick and dirty way of checking that the steps completed
without error, and checking that their corresponding qsub jobs
completed. Since splitting reads by sample and tagcleaning are
qsubbed, it is possible for the Part1 to hang if the jobs
complete with neither errors nor the expected output.

Each sanity check is based on counting lines of the input and
output files.
The number of lines in barcodes.fastq should equal the number of
lines in the raw index files. The sum of lines in sample-specific
fastqs should equal the number of lines in each seqs.fastq.
Tagcleaned files must have the same number of lines as their
corresponding un-tagcleaned files.

Bugfix: Request 400M unallocated RAM for tagcleaner.pl jobs.