Part1.sh -d tagclean requires split-by-sample fastqs only (v2.1.2-allow_intermediates+1)
Started this branch as a step on the path toward making the
pipeline more of a toolbox. This branch will cause -d to remove
dependency on the raw files and any intermediates not relevant to
the selected pipeline step. A future branch will use bash-style
options to provide access to the individual steps, each with
their own set of command-line options. The whole pipeline with
full checking for semantic errors will continue to be the default.

-mapping file validation is now a part of the barcode extraction
 step. This means that map validation occurs on every pipeline run
 without -d or with -d barcodes, but can be skipped with -d demux,
 -d tagclean, or -d dada2.
-Existence of barcodes.fastq does not cause the pipeline to skip
 map validation.