Stop if DADA2 stats malformed. Receive zip archives from part1 (v2.3.0-report+2)
When making the report in part2, if non-interactive, then stop
for any lines in the DADA2 stats file without 4 tab-delimited
fields. If interactive, allow the user to choose a the DADA2
stats file manually. Lines without a tab will simply be taken
as the end of the file.

When generating the report, expect the fastQC reports as zip
archives.

Print the run name (not the path of the run on the grid) as the
header for FastQC reports.

Show "Loading..." while the report is loading in a browser.