Qsub illumina_dada2.pl with hard requirement for 4G mem_free (v2.7.0-part1_4g+1) (v2.7.1)
Patch applied in response to a crash of print_qiime_config.py.
Mem_free requirement was expanded from the miserly 200M
to a generous 4G. In the future, make an effort to utilize the
same node for tasks that are currently qsubbed by
illumina_dada2.pl.

Squash-merge-committed from patch/part1_4g