Override taxon labels after renaming ASVs (v2.7.0-custom_sequence_labeling+1)
Renaming ASVs silently broke part2's ability to override the
classification of four ASV sequences, which are known to be Candida
albicans, BVAB1, and BVAB_TM7. Functionality is restored by
reading the *ASV.fasta reference file in the project directory,
searching for appearance of any of the four sequences, and
replacing the taxon assigned to the corresponding ASV in the
in-memory dictionary.