Only custom-label ASVs for SILVA and HOMD (v2.7.0-custom_sequence_labeling+2)
Custom-labeling ASVs is not thought to be helpful for when
sequencing ITS with UNITE as the taxonomic reference, as BVAB1
and TM7_BVAB are bacterial, and UNITE has reference sequence(s)
for Candida albicans.

Also, streamlined control flow for so count tables for the
individual taxonomic references used are always produced.
(Previously, running with -v V3V4 wouldn't be sufficient to
output SILVA classification tables, even though assignment from
SILVA occurred.) Now --pecan-silva is ignored, although I left
it in for compatibility.