Override taxon labels after renaming ASVs (v2.8.1)
Renaming ASVs silently broke part2's ability to override the
classification of five ASV sequences, which are known to be Candida
albicans, BVAB1 (x3), and BVAB_TM7. Candida albicans is fungal and
contained in UNITE but not SILVA, yet is relevant to vaginal and
other health areas. BVAB1 and BVAB_TM7 still have murky
classification and are thus not yet in SILVA.

Functionality is restored by reading the *ASV.fasta reference file
in the project directory, searching for appearance of any of the
five sequences, and replacing the taxon assigned to the
corresponding ASV in the in-memory dictionary. This is done for
SILVA, HOMD, and PECAN abundance tables, leaving the classification
FASTA/TXT file as it is.

This update also makes the production of count tables more
intuitive. (Whenever targeting V3V4, both SILVA and PECAN count
tables, plus the SILVA+PECAN count table, are produced.)


Former-commit-id: 9c2540e960a06c7bc116036893ab3c875c08d85d