Offer --troubleshoot_barcodes and allow numeric sample names (v2.10.0-troubleshoot_barcodes+1)
--troubleshoot_barcodes causes the pipeline to try all
permutations of reverse-complementing and switching indexes. The
pipeline will create additional temporary directories in the
project directory to extract alternative versions of
barcodes.fastq. If any samples are demuxed, all created files
will be copied back to the original run directory, and the
pipeline will continue. Without --troubleshoot_barcodes, the
pipeline dies if no samples are demuxed.

A bug in the report script caused it to crash because sample
names would be parsed by NumPy as floats not strings.