Stop after barcode troubleshooting; don't try writing to repo (v2.11.4)
- Try demuxing using all combinations of indexes. Recently, we've
seen projects where most combinations of indexes yield a few reads
demuxing. This makes it difficult to automatically choose the best
index configuration. Now, the pipeline stops after trying all
configurations, and leave it to the user to inspect the files,
move the correct ones to the run directory, and continue the
pipeline.

Bugfix:
- While creating the HTML report, don't attempt modifying files
in the repo, because most users don't have sufficient permissions.
It is safe to just write files to the report directory instead of
writing them to the repo and copying them over.