Provide a dada2 workflow for PacBio CCS reads (v2.9.1-pacbio+15)
Squash-merged from feature/pacbio

New pacbio_dada2.R script that is roughly analogous to
illumina_dada2.pl. The new script produces the same files (and
some QC visualizations), but has a different CLI and needs to
be run in the run directory in a qlogin session.

The ./taxonomy/ subdirectory and its large files has been eliminated
in favor of ./config.json, which allows a user to specify the path
to taxonomy reference databases.

The consolidation of run data and subsequent removal of bimeras
has been relegated to a standalone R script.

Taxonomy assignment has been relegated to a single standalone R
script with an interface that accomodates all of the previously
used reference databases.

Lastly, some scripts have been moved from personal directories to
the ./scripts/ directory.


Former-commit-id: 0febc1b727ec67ea6d841f04ebbe95de17641bb5