Add part3_cleanup.py and its synology client module (v2.12.1-tweaks_for_cleanup+2)
part3_cleanup.py - an interactive script that reorganizes the
results of part1 and part2 to prepare the project for delivery.
Also enables the user to upload files to JIRA and a Synology drive.

The Synology client module was written to provide an interactive
access to Synology, but also has functions imported by part3.