Write platform to metadata file during part1
Part1 now stores checkpoints in .meta.json. That file's JSON
contents are divided into a 'checkpoints' object and a 'params'
object. 'params' is used to record whether the platform is
ILLUMINA or PACBIO. part3_cleanup.py reads this parameter to
determine which citations to include.

During part1 for either Illumina