Use software installed on IGS RHEL8 (v3.0.2)
Added config entries for local QIIME binaries, R 4.0 and 3.6
binaries and libraries, 16S databases, and python2. Updated paths
in some shebangs (is this really the right way to get paths to R
and perl binaries and libraries?). A future patch will put all
paths in config.json, so the pipeline is more portable.

Part1:
- Repaired pass-through of DADA2 truncation length parameters.
- Put DADA2 QC and denoising into separate script.
- Changed the required output of DADA2 step from artifacts to
  actual data files

R scripts: switched from using rjson to jsonlite, because
jsonlite is installed for all users, and must be loaded before
finding a custom package library.

Patched make_report.py to not crash when samples identified as
controls in the map have 0 reads surviving after DADA2.

commit 8de2eaf2c7ab584463a86e38b8bc98f7c0b609a0
Author: Jon Lim <jlim7@comcast.net>
Date:   Wed Apr 21 14:23:50 2021 -0400

    Updated part1.sh for RHEL8 (v3.0.1-RHEL8_backwards_compatible+1)

    When IGS upgraded to RHEL8, usepackage was replaced with
    environment modules. Part1.sh has been updated accordingly for
    compatibility with either system.