Call python scripts with -s option to interpreter (v3.0.2-RHEL8_backwards_compatible+3)
To use qiime, it is not necessary to activate the qiime1 env.
Instead, we put the path to the qiime1 env binaries in the
pipeline config file, and use it to find the python interpreter
and the qiime scripts.

Whenever a python script is run, we prevent any influence from
the 'site' module by running:

  <python_2_or_3> -s script.py -option1 -option2 ...

The site module adds user-specific "installation-dependent" paths
to the module search path. We also set PYTHON_PATH to "" at the
very start of part1 and part2.