Give more convenient filenames for sample FASTQ's (v3.0.6)
Squashed merge of user_friendly_run_combining

When the project is ready for delivery, raw and adapter-trimmed
FASTQ's should have the same name as what is in the count tables,
<RUN>.<SAMPLE NAME> with the _R1 or _R2 suffix indicating read
orientation. This naming scheme involves modifying the sample
names in the map and adding suffixes to the FASTQ's immediately
after the demultiplex.

Misc:

Part 1 wrapper
- Warn of potential SGE errors when submitting the job

Part 2
- Fix file open mode error

Part 3
- At the start of preparing files for delivery, if user doesn't
  select any runs, ask them to confirm. This is because it's easy
  to assume the Enter key selects a run, when instead it just
  confirms the current selection.
- Fixed typo in read_metadata()
- Fixed so messages regarding the MAP directory is consistent with
  other messages in showing relative path.
- When organizing for archival; if any unexpected excpetions,
      print and ask user if they want to continue