In part 2, use provided project map to control run merging (v3.0.7-control_run_merging+1)
The project map (by default, project_map.txt) is a tab-delimited
file with two columns, "RUN.PLATEPOSITION" and "sampleID". It
ensures that the desired samples are imported into the project.

The rationale is that recently, we demux all samples in each run,
using a run-agnostic QIIME map. As a side-effect of this new
process, the project map will be less prone to error than the
QIIME map.

The QIIME maps are now concatenated to a file called qiime_map.txt