Use project map to select samples from runs during part 2 (v3.1.0) (v3.1.0)
Squashed commit of the following from feature/control_run_merging:

commit 24a4f357ecf3f9440a57f7f0f86ade7478acb813 [formerly a01307032acc217968e51f2f0a09b547b8fa3977]
commit c91ea3fab8daed46a2e9401ba45ea5df4c91cdd3 [formerly 714cd3e236576a3c5406bb50ca95d8a83603d618]
commit 61f3a55d6e01e77240d2f04d6eda3751c314c58b [formerly f38d35f8fc687b355d00f68225b1f749f204c884]
commit 9baad0fdad805941a8563290c9cfbeec8f15b106 [formerly d09a1a65e13091cd66dd6687e47178d09435bdee]
commit a7fd0b03345c04d22b8e57288a5be818772e3ea2 [formerly e61fb55656f2a1516018f68f4361c38111fe8f91]
commit 953449dc48f2d8bcbff5b590a64969449162b236 [formerly 68d3a95b3080c18f0a3da93b8666aaefb174a968]

In part 2, the new optional parameter --map takes a map file that
tells the pipeline which samples to include. The map is a tab-
delimited file with exaclty two columns: RUN.PLATEPOSITION and
sampleID. The column RUN.PLATEPOSITION should match the sample
names output by part 1. The column sampleID prescribes the
samples' new (ie, researcher-assigned) names.

In addition, when any values in RUN.PLATEPOSITION resemble UDI
identifiers, the controls run on the same plates are automatically
included. ./share/controls_platepositions.csv is used to find
controls, since controls are expected to be in the same plate
positions for every UDI run. If the project map contains a sample
in the position where a control is expected, the sampleID
in the map will not be overridden by the ID of the control.

When --map is absent, part 2 will simply include all smples in
the specified runs.

Accordingly, during part 3, the user must specify which file is
the map, befoer uploading it to JIRA.


Former-commit-id: cbd0ddf4d4c2fb2d84e0917682f991482b54eab4