Part 1 + Part 2: Use MSL T3 directory for runs by default (v3.1.1-configure_run_storage+1)
Added config variable 'run_storage_path' giving the default place
for run storage.

Part 1: The wrapper script for Illumina runs uses
'run_storage_path' as the default location to place the run
directory. If that locaiton alreaady holds a directory with the
same name as the run, the user is asked to confirm before
overwriting.

Part 2: 'run_storage_path' is used as the default location for
finding runs. Runs are accepted from the option -r; the option -i
is now deprecated. Run names are required to be the names of the
run directories; anything that looks like a path will not be
accepted.

For both parts, the option --run-storage can be used to change the
run storage path.