In part 3, pass runs from part 2 to part 3 (v3.1.1-configure_run_storage+2)
In part 2, symlink each run directory to the project directory if
it is not a subdirectory of the project directory.

In Part 3, show all subdirectories of the project directory and
ask the user to select runs. When showing this checkbox interface,
pre-select the items that match the names of runs in the project
metadata.

Also, expanded the bash wrapper for part 1 for PacBio runs, so
it also uses the global run storage location by default. This
wrapper can be submitted to the grid from anywhere, but it is still
suggested to be run interactively, since the wrapper will ask to
confirm overwriting an existing run.