Don't delete symlinked runs in part 3 (v3.1.1-configure_run_storage+3)
In Part 3, the user is allowed to choose runs from the project
directory, including runs that were symlinked there by Part 2.
When Part 3 is cleaning up the project directory, the symlink
should be deleted, but not the linked directory.

Also fixed an uncaught exception when Part 3 tries to change
permissions on a broken symlink.