Fix bug in getting citation for SILVA databases (v3.1.1-configure_run_storage+4)
When running part 2 targeting V3V4 or V4, the taxonomic annotation
tables need to be named after the specific version of SILVA, so
the correct annotation can be added. It is now Part 2's
responsibility to determine what version of SILVA to use, when the
version is not explicitly specified.