Use MSL T3 directory for runs by default  (v4.0.0)
Added config variable 'run_storage_path' giving the default place
for run storage.

Part 1: The wrapper script for Illumina runs uses
'run_storage_path' as the default location to place the run
directory. If that location alreaady holds a directory with the
same name as the run, the user is asked to confirm before
overwriting.

Part 1 for PacBio also uses the global run storage location
by default. This wrapper can be submitted to the grid from
anywhere, but it is still suggested to be run interactively,
since the wrapper will ask to confirm overwriting an existing run.

Part 2: 'run_storage_path' is used as the default location for
finding runs. Runs are accepted from the option -r; the option -i
is now deprecated. Run names are required to be the names of the
run directories; anything that looks like a path will not be
accepted.

At the end of Part 2, any runs that are not in the project
directory are symlinked into the project directory so they can be
selected in Part 3. In part 3, check metadata from part 2 to pre-
select runs from among the subdirectories of the project.

For both parts, the option --run-storage can be used to change the
run storage path.

Usability:

Enabled --option=value syntax for most of the long-form options
on the part 1 bash wrappers.

Barcode troubleshooting message (at the end of troubleshooting)
is now printed to stdout and the log.

Made the remove_bimeras.R produce the full count table containing
samples from all runs, before applying the project map. This
enables the user to troubleshoot map problems.

Simplified taxonomic annotation scheme selection
The base taxonomic annotation scheme is now determined by the variable
region. Additional taxonomic annotation schemes can be run using
--tax. --pacbio works by setting the default region to FULL-LENGTH.
--homd works by setting the annotation scheme to HOMD, regardless
of variable region.

Moved taxonomy database filepaths to the pipeline config file.

Bugfixes:

Fixed bug in getting citation for SILVA databases.
When running part 2 targeting V3V4 or V4, the taxonomic annotation
tables need to be named after the specific version of SILVA, so
the correct annotation can be added. It is now Part 2's
responsibility to determine what version of SILVA to use, when the
version is not explicitly specified.

Fixed bug where taxonomic annotation files wouldn't be tracked
when multiple annotation schemes were called.

Added citations for SILVA132 and SILVA138forPB to Part 3

Squashed commit of the following merged from version/configure_run_storage

commit 454e0d81b24fba7320b3d2b7e59a58c56b31a256
commit e595938faaf97d5ef988cc5d0addf97c095e27a3
commit 223a791d01cdf769c2c1d268c02f34ba71748a94
commit 56e8bb4df971331270cfca5fc6222e780cbf71c3
commit 1f37d73fdf0c0df059e00ae5329e26a8876a887a
commit abe0af93d6c14805fed57a4543f11c08ca44b079
commit 87c40b3b6565d638175ace3bf9e9bde4584324c4