Make any created files and directories in T3/MSL shared (v4.0.0-share_files_in_T3+1)
In Part 1's wrappers and perl scripts, if the run directory
is a subdirectory of the global run directory, apply umask 0002
so all created files and directories are shared with the group.

I think all pipeline operators are in group 'igs' by default, but
it would be nice to just ensure that everything in the global
run directory is of the same group. This can be done by executing
chmod g+s on the global run directory to apply the setgid bit.