In Part 3, qsub long filesystem operations (v4.2.0)
Squashed merge from feature/part3_qsub

When copying raw reads to the project directory, part 3 now gives
users the option to qsub the job. This enables it to run in the
background.

As a side effect, I've added most of the base command for qsub,
weth the most commonly used options, to the pipeline config. This
is used in all part 1 wrappers and scripts.

Maintainability:
Condense code for adapter trimming in part 1 for Illumina

Squashed from the following:

commit d91bdcfa9b5e325686c96207d81bfa24c1db21c0
commit f178b9f0c98ddf3b062f801c9bcdf8074fdc928f
commit 26fae46f8f48632a6e4fa17b9af4f4a6bbccc457