Share all files made in part 1 whenever the script finishes (v4.2.0-share_files_in_T3+4)
For the Illumina part 1, it's necessary to give group permissions
to new files after making them, because they are made in qsubbed
jobs, and I don't want to prepend 'umask 0002' to every qsubbed
command. An alternative short-term solution is to run all the
commands in serial on the same shell (which is qsubbed already
by part1.sh). Another alternative is to create a bash script
formatted for the SGE for each command, and submit that.