Share part 1 Illumina runs in the shared run directory (v4.2.2)
Squashed merge from patch/share_files_in_T3

For the Illumina part 1, it's necessary to give group permissions
to new files after making them, because they are made in qsubbed
jobs, and I don't want to prepend 'umask 0002' to every qsubbed
command. An alternative short-term solution is to run all the
commands in serial on the same shell (which is qsubbed already
by part1.sh). Another alternative is to create a bash script
formatted for the SGE for each command, and submit that.

Changes to branch since the last merge:

commit 5521f751dc0a02e1585a2a2c20f26eb4d145cc11
commit 4fbf23174bbc7020b2b429b884fa699fa9a5f381
commit e6b8e718ed8052393631353b9bf6e982de4c9a8f