Update UNITE db and main script documentation (v4.2.3-update_UNITE_db_and_main_script_documentation+1)
UNITE updated from 2017/12 release to 2021/05 release. There are
two UNITE releases: the general release and the developer release.
The general release is pre-trimmed to the ITS region, but the QIIME2
developers say the full sequences are better (and I don't know if
our primers flank the same region as the general release), so here
I took the dev release.

(https://docs.qiime2.org/2020.2/tutorials/feature-classifier/)

Also, I took the more stringent release including singletons only
if they are RefS (human-vetted). Compared to the 2017 release,
this is still an improvement because while there are roughly the
same number of sequences, more of them are RefS than in the 2017
release.

Also updated documentation for:
- part1.sh
- part1_pacbio.sh
- illumina_dada2.pl
- pacbio_dada2.R