Increase tagcleaner sensitivity to 7 mismatches (v4.3.0-primer_trimming_qc+2)
Squashed commit of the following:

commit b7091f49fa63c1efcd2a95a6acc0159ce85b244c

This is a stopgap measure to trim primers more aggressively,
decreasing the likelihood that an untrimmed ASV will appear in
the DADA2 output. Will be replaced soon with a solution for
removing untrimmed reads.

Also sneaking in a hotfix for more stringent pattern matching in
part 1 when finding primer-trimmed files.


Former-commit-id: 928df94a06b38dae3f2c218fcca093d1da401a8e