Ensure all primer trimming finishes before continuing (v4.3.2)
We don't have an accurate way of determining when jobs submitted
to the SGE are finished, from perl. Tagcleaner makes its output
file and then appends to it, so we need to find a different way
of determining when it's done. For now, just run all tagcleaner on
the same node as the pipeline.


Former-commit-id: c7458ee8de90bda891c3c97c397449c3afb33de7