Use perl module to track status of primer trimming jobs on SGE (v4.3.1-primer_trimming_qc+1)
Adapted SGE::Control to use qstat to check job status. Tried but
failed to install DRMAA due to lack of some library or header.
When running a command, we just use backticks, put the stdout
and stderr to a file, and read the file to get the job ID. Then,
we call qstat and parse the output to see what jobs are still
running.

So now, it's safe to just call execute_and_log(..qsub=1) and
move on with the pipeline.