Get controls from all runs (v4.3.3-get_controls_from_all_runs+1)
Previously, we only got one of each control for each UDI plate in
the project; which run the controls came from was undetermined.
Now, for each UDI plate, we determine which runs had that plate,
then add lines to the map to include the controls from those
runs. This does result in ambiguous control naming; the operator
should designate unique names for the controls in the project map,
if desired.