Correct V4 reverse primer and Schedule::SGE improvements (v4.3.6)
V4 reverse primer was previously same as the V3V4 fwd primer. Changed it
so it's the same as the V3V4 rev primer.

If all fwd or rev reads were filtered out when trimming primers
from a sample, say whether it was fwd or rev.

Prevent stderr message when reading qstat when a node's load_avg
is "-NA".
Echo the Job ID to STDOUT when calling Schedule::SGE::execute().

Squashed commit of patch/fix_v4_rev_primer:

commit 3ed7fbeead4bab2aba3502b0164ef8bf5a07cc71
commit d061e7523e04974a7cae2140fb33c1e127008a1e