Specify python2 interpreter to Python2 scripts (v4.3.8)
In part 2, rename_asvs.py was written in Python2. Instead of
trying to ensure Python2 is on the path when the script is called,
just feed it to the Python2 interpreter specified in pipeline config.

(This follows my general trend of not relying on shebang lines,
since they rely on heavily on the user environment)

Squashed commit of patch/dont_use_python2_shebang:

commit 6b71f39d3b50b1d0ab7ad784f8888453799447a3