Get stats even if no split library log (v4.4.2)
Improvements to running Illumina part 1 on demuxed files:
- If the split library log doesn't exist, count the reads in the
  demuxed files for the 'Input' column of dada2_part1_stats.txt.
  If unable to do so, just put NA in that column.
- Improved part1.sh documentation to make it easier to understand
  filename requirements when starting with demuxed files.

Fixed bug in part1.sh that causes the script to hang when trying
to parse certain options with unrecognized values.

For the Illumina DADA2 script, stopped producing .Rout file (which
is an artifact from using R CMD BATCH, which we no longer do). R
messages now go to STDERR (like all R messages) and the pipeline
log points to the R STDERR when an error is detected.


Former-commit-id: 7255b5376950e61cedc51290b6bcca290352936c