Make project directories chmod 2770, update UNITE db (v5.0.1)
Uses unix `install` to make project directories owned by group igs
and with perms 2770. The 2 sets gid bit so anything created inside
the directory to also be owned by group igs. In future, please
isolate this code to a wrapper or a separate file of parameters
specific to MD Genomics.

Uses the 2022.11.29 version of UNITE DB

Get both the speciateIT output file and the cleaned version of the
file during part3.
Updated _CITATIONS.txt placed in project directories by part3.

Squashed commit of the following:

commit 9a86eb8a794e0c3e329b248e01607803f7230d5d