Rework project/sample informatics (v6.0.0-accept_xls_platemaps+1)
First join manifest to plate maps, if available. Sometimes the manifest itself
  will serve as the platemap (rename the columns). Sometimes we receive beads in
  plates. We extract gDNA and add controls; this means the manfiest sample names
  are the final sample names (they trump conflicting sample ID's on the gDNA
  maps) but gDNA map samples NOT on the manifest will get added to the project
  too.
Then join this sample map to the pooling detail. This enables to create the
  names of files and data table rows that will be grabbed from run demux
  directories.
The "Sample Description" column of the pooling detail can be used to identify
  the plate, if gDNA plate ID is N/A. This reflects that at pooling, plates can
  come from two different sources (they're two different types of plates).
  I treat them identically, but not everyone does.
Do some CLI argument validation at the very start.

Keep the APJR3 barcode map in the repo.